Upgraded NavBar Transition for Small and Medium Sized Devices
Perviously, the vertical navbar on small and medium sized devices was not fluid, felt jerky and buggy. I fixed this by adding a fade in on the background blur, as well as a smaller time delay, and ease on transition. Changes made to JS code and 'close_bg' class. Additonally, a new class of 'visible_close_bg' was added to aid the transition.