Completed Blog Page
The main change in this commit was the completion of the Blog webpage for the website. This page includes a content section, side bar, and comment box. Nothing on the page has actual functionality yet (I will consider adding a posting mechanism for comments) but the UI is completed. Additionally I increased responsiveness of the website on the sub-pages (About, Blog, Courses, Contact).