Completed Contact Page
Completed the UI for the 'Contact' webpage for the website. Additionally, improved the method by which horizontal scrolling (caused by closed vertical nav bar in mobile view) was prevented. This was by giving the vertical nav bar a 'position: fixed' property instead of preventing overflow inside the nav bar container.