PHP Written for Contact Form Submission
Created 'formHandler.php' that allows for contact form to be submitted and an email to be sent out to myself with the message of the form, as well as, a confirmation email to be sent out to the form submitter to let them know their form was submitted. This is not live yet and still needs to be tested thoroughly.

Signed-off-by: Muhammad Ahmed <moahzed@gmail.com>